+++ src/controllers/GraphqlController.php
@@ -156,6 +156,13 @@ class GraphqlController extends Controller
         // Generate all transforms immediately
         Craft::$app->getConfig()->getGeneral()->generateTransformsBeforePageLoad = true;
 
+        // Check for the cache-bust header
+        $gqlCacheHeader = $this->request->getHeaders()->get('x-craft-gql-cache', null, true);
+        if ($gqlCacheHeader === 'no-cache') {
+            $cacheSetting = Craft::$app->getConfig()->getGeneral()->enableGraphQlCaching;
+            Craft::$app->getConfig()->getGeneral()->enableGraphQlCaching = false;
+        }
+
         $result = [];
         foreach ($queries as $key => list($query, $variables, $operationName)) {
             try {
@@ -175,6 +182,10 @@ class GraphqlController extends Controller
                     ],
                 ];
             }
+        }
+
+        if ($gqlCacheHeader === 'no-cache') {
+            Craft::$app->getConfig()->getGeneral()->enableGraphQlCaching = $cacheSetting;
         }
 
         return $this->asJson($singleQuery ? reset($result) : $result);